Commented out report.dataItems lines because we can get dataItems through Query objects. No point in scanning the XML multiple times for the same information.
Modified HTMLify() to go down the hierarchy from Report, through Query to DataItem.
This way, we don't have to send JSON to the method and have more fine control over the data inside the method.

Added simple JavaScript to template for showing/hiding sections. This can be done with JQuery or something later, but I just wanted to get the structure working.